+++ probe-rs/examples/benchmark.rs
@@ -1,3 +1,4 @@
+use probe_rs::Permissions;
 use probe_rs::{config::TargetSelector, MemoryInterface, Probe, WireProtocol};
 
 use std::{env, num::ParseIntError, time::SystemTime};
@@ -6,20 +7,21 @@ use std::{
     time::{Duration, Instant, UNIX_EPOCH},
 };
 
+use clap;
+use clap::Parser;
 use rand::prelude::*;
-use structopt::StructOpt;
 
-#[derive(StructOpt)]
+#[derive(clap::Parser)]
 struct Cli {
-    #[structopt(long = "chip")]
+    #[clap(long = "chip")]
     chip: Option<String>,
-    #[structopt(long = "address", parse(try_from_str = parse_hex))]
+    #[clap(long = "address", parse(try_from_str = parse_hex))]
     address: u32,
-    #[structopt(long = "speed")]
+    #[clap(long = "speed")]
     speed: Option<u32>,
-    #[structopt(long = "protocol")]
+    #[clap(long = "protocol")]
     protocol: Option<String>,
-    #[structopt(long = "pr")]
+    #[clap(long = "pr")]
     pr: Option<u64>,
 }
 
@@ -32,7 +34,7 @@ const SIZE: usize = 0x1000;
 fn main() -> Result<(), &'static str> {
     pretty_env_logger::init();
 
-    let matches = Cli::from_args();
+    let matches = Cli::parse();
 
     let mut probe = open_probe(None)?;
 
@@ -69,7 +71,7 @@ fn main() -> Result<(), &'static str> {
     let probe_name = probe.get_name();
 
     let mut session = probe
-        .attach(target_selector)
+        .attach(target_selector, Permissions::default())
         .map_err(|_| "Failed to attach probe to target")?;
 
     let chip_name = session.target().name.clone();