+++ probe-rs/src/config/target.rs
@@ -2,6 +2,7 @@ use probe_rs_target::{Architecture, ChipFamily};
 
 use super::{Core, MemoryRegion, RawFlashAlgorithm, RegistryError, TargetDescriptionSource};
 
+use crate::architecture::arm::sequences::nrf53::Nrf5340;
 use crate::architecture::arm::sequences::nxp::LPC55S69;
 use crate::architecture::arm::sequences::ArmDebugSequence;
 use crate::architecture::riscv::sequences::esp32c3::ESP32C3;
@@ -89,12 +90,15 @@ impl Target {
             Architecture::Riscv => DebugSequence::Riscv(DefaultRiscvSequence::create()),
         };
 
-        if chip.name.starts_with("LPC55S69") {
-            log::warn!("Using custom sequence for LPC55S69");
+        if chip.name.starts_with("LPC55S16") || chip.name.starts_with("LPC55S69") {
+            log::warn!("Using custom sequence for LPC55S16/LPC55S69");
             debug_sequence = DebugSequence::Arm(LPC55S69::create());
         } else if chip.name.starts_with("esp32c3") {
             log::warn!("Using custom sequence for ESP32c3");
             debug_sequence = DebugSequence::Riscv(ESP32C3::create());
+        } else if chip.name.starts_with("nRF5340") {
+            log::warn!("Using custom sequence for nRF5340");
+            debug_sequence = DebugSequence::Arm(Nrf5340::create());
         }
 
         Ok(Target {